delete with validation

+++ src/items/items.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateItemDto } from './dto/create-item.dto';
 import { UpdateItemDto } from './dto/update-item.dto';
 import { Repository } from 'typeorm';
@@ -31,7 +31,12 @@ export class ItemsService {
     })
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const item = await this.itemRepository.findOneBy({ id })
+    if (!item) {
+      throw new NotFoundException(`Not found: id=${id}`)
+    }
+    
     return this.itemRepository.delete({ id }) 
   }
 }
